pyproject.toml
[project]
name = "analyst-robot"
version = "0.1.0"
description = "ai assistant for stock analysis"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "ipykernel>=6.29.5",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-openai>=0.3.17",
    "langgraph>=0.4.5",
    "numpy>=2.2.5",
    "openai>=1.78.1",
    "pandas>=2.2.3",
    "pytest>=8.4.0",
    "pyyaml>=6.0.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",  
]

[tool.pyright]
include = ["src", "main.py"]
reportMissingImports = "warning"
reportUnknownParameterType = "warning"

[tool.pyright.executionEnvironments]
root = "."

[project.optional-dependencies]
dev = [
    "ruff>=0.5.0",
]

[tool.ruff]
line-length = 120
target-version = "py312"
include = ["src"]
